+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Fix Mirror of Night "locked" bottom panel when in `vert+`.
+-   Fix locked choice overlay (Approval Process pact) when in `vert+`.
 
 ## [1.7.0] - 2022-07-08